tidy add IngestDocCleanupMixin
